pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bedrock-forge"
version = "0.1.0"
description = "A unified Python CLI for orchestrating Bedrock-based WordPress workflows"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "Bedrock Forge Team", email = "team@bedrock-forge.com"}
]
maintainers = [
    {name = "Bedrock Forge Team", email = "team@bedrock-forge.com"}
]
keywords = ["wordpress", "bedrock", "cli", "deployment", "devops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
    "Topic :: System :: Software Distribution",
    "Topic :: System :: Systems Administration",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
]
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.17.0",
    "rich>=14.0.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "SQLAlchemy>=2.0.0",
    "redis>=6.0.0",
    "aiohttp>=3.10.0",
    "httpx>=0.25.0",
    "requests>=2.30.0",
    "hcloud>=2.0.0",
    "cloudflare>=4.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.20.0",
    "google-auth-oauthlib>=1.0.0",
    "paramiko>=4.0.0",
    "PyNaCl>=1.5.0",
    "celery>=5.0.0",
    "structlog>=25.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
    "tqdm>=4.60.0",
    "psutil>=6.0.0",
    "packaging>=25.0.0",
    "python-whois>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.0.0",
    "coverage>=7.0.0",
    "black>=24.0.0",
    "isort>=5.12.0",
    "flake8>=7.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/bedrock-forge/bedrock-forge"
Documentation = "https://docs.bedrock-forge.com"
Repository = "https://github.com/bedrock-forge/bedrock-forge.git"
"Bug Tracker" = "https://github.com/bedrock-forge/bedrock-forge/issues"
Changelog = "https://github.com/bedrock-forge/bedrock-forge/blob/main/CHANGELOG.md"

[project.scripts]
forge = "forge.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["forge*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
forge = ["*.json", "*.yaml", "*.yml", "config/*", "templates/*"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["forge"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["forge/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["forge"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]